Select Favorite and database working correct. FavoriteScreen works as intended

+++ composeApp/src/commonMain/kotlin/com/plcoding/bookpedia/book/data/dto/SearchedBookDto.kt
@@ -16,6 +16,6 @@ data class SearchedBookDto(
     @SerialName("first_publish_year") val firstPublishYear: Int? = null,
     @SerialName("ratings_average") val ratingsAverage: Double? = null,
     @SerialName("number_of_pages_median") val numPagesMedian: Int? = null,
-    @SerialName("ratings_count") val ratingsCount: Int? = null,
+    @SerialName("ratings_count") val ratingsCount: Double? = null,
     @SerialName("edition_count") val numEditions: Int? = null,
     )
